requirements.txt
Django>=5.2.6,<5.3
djangorestframework>=3.16.1,<3.17
psycopg[c]>=3.2.10,<3.3
drf-spectacular>=0.28.0,<0.29
django-filter>=25.2,<26